Add extra info bar to display page to show upload ID and view raw button
Browsers do not display the URL in an unambiguous font, so it can be
hard to read the upload id by eye. Add a second bar so the id is
displayed in a monospace font in the page body. Also make the filename
monospace font so it is easy to read too. Two bars gives more space, so
add a "View raw" button which is good for copy-pasting the entire file.