web: Ignore clippy::result_large_err on views::render_template
It's angry that the ErrorResponse is too large (> 128 bytes). This isn't
a problem because it actually is just a wrapper around Response, the Ok
value, so the Ok and Err are both "large". This also isn't public API so
there isn't the risk of infecting someone downstream with a large error
value.